delete last message in clear cash if has forward or replay

+++ app/src/main/java/net/iGap/realm/RealmRoomMessage.java
@@ -460,11 +460,11 @@ public class RealmRoomMessage extends RealmObject {
                 public void execute(Realm realm) {
                     RealmQuery<RealmRoomMessage> roomRealmQuery = realm.where(RealmRoomMessage.class);
                     for (RealmRoom realmRoom : realm.where(RealmRoom.class).findAll()) {
-                        if (realmRoom.getLastMessage() != null) {
+                        if (realmRoom.getLastMessage() != null && realmRoom.getLastMessage().getForwardMessage() == null && realmRoom.getLastMessage().getReplyTo() == null) {
                             roomRealmQuery.notEqualTo(RealmRoomMessageFields.MESSAGE_ID, realmRoom.getLastMessage().getMessageId());
                         }
 
-                        if (realmRoom.getFirstUnreadMessage() != null) {
+                        if (realmRoom.getFirstUnreadMessage() != null && realmRoom.getFirstUnreadMessage().getForwardMessage() == null && realmRoom.getFirstUnreadMessage().getReplyTo() == null) {
                             roomRealmQuery.notEqualTo(RealmRoomMessageFields.MESSAGE_ID, realmRoom.getFirstUnreadMessage().getMessageId());
                         }
                     }
@@ -478,7 +478,7 @@ public class RealmRoomMessage extends RealmObject {
 
                     RealmRoom realmRoom = realm.where(RealmRoom.class).equalTo(RealmRoomFields.ID, roomId).findFirst();
                     if (realmRoom != null) {
-                        if (realmRoom.getLastMessage() != null) {
+                        if (realmRoom.getLastMessage() != null && realmRoom.getLastMessage().getForwardMessage() == null && realmRoom.getLastMessage().getReplyTo() == null) {
                             realm.where(RealmRoomMessage.class).equalTo(RealmRoomMessageFields.ROOM_ID, roomId).notEqualTo(RealmRoomMessageFields.MESSAGE_ID, realmRoom.getLastMessage().getMessageId()).findAll().deleteAllFromRealm();
                         } else {
                             realm.where(RealmRoomMessage.class).equalTo(RealmRoomMessageFields.ROOM_ID, roomId).findAll().deleteAllFromRealm();
